+++ CHANGELOG.md
@@ -180,6 +180,7 @@
 - PR #5184 Fix style checks
 - PR #5198 Add detail headers for strings converter functions
 - PR #5199 Add index support in `DataFrame.query`
+- PR #5209 Add `nan_as_null` support to `cudf.from_pandas`
 - PR #5207 Break up backref_re.cu into multiple source files to improve compile time
 - PR #5214 Move docs build script into repository
 - PR #5155 Fix cudf documentation misspellings